// src/app/interfaces/book.interface.ts
export class BookInterface {
    _id?:                string;
    ISBN:                string | number;
    Book_Title:          string | number;
    Book_Author:         string | number;
    Year_Of_Publication: string | number;
    Publisher:           string | number;
    Image_URL_S:         string | number;
    Image_URL_M:         string | number;
    Image_URL_L?:        string | number;
}